chore: update full width container for product pages
