+++ crowdin/messages.json
@@ -1112,7 +1112,6 @@
   "1542393407": "CFD trading features on Deriv",
   "1546535785": "Are there any conditions that I should meet before I can withdraw my commission?",
   "1546917177": "5 - 750",
-  "1549497379": "High leverage, tight forex spreads",
   "1550284135": "Jordan",
   "1550574351": "Let’s say you want to keep two lots of EUR/USD with a point value of 0.00001 and swap rate of -0.12 open for one night.",
   "1551153165": "You can trade the following options on Deriv:",
@@ -2360,11 +2359,11 @@
   "-327214682": "25+ crypto pairs",
   "-819550264": "Zero commission",
   "-601411843": "Derived FX gives you the opportunity to trade on simulated assets based on major forex pairs at the volatility of your choice. Our advanced algorithms track real-world currency prices and dampen the fluctuations caused by market sentiment and unexpected news events.",
-  "-1389465070": "Why trade Derived FX on Deriv",
   "-895517965": "Derived FX trades available on Deriv",
+  "-1389465070": "Why trade Derived FX on Deriv",
   "-393616618": "Start trading Derived FX on Deriv in 3 simple steps",
-  "-1379291897": "Why trade forex on Deriv",
   "-1573631881": "forex",
+  "-1379291897": "Why trade forex on Deriv",
   "-673420280": "Start trading forex on Deriv in 3 simple steps",
   "-1718265662": "Open a real account, make a deposit, and start trading stocks, indices and other markets.",
   "-1823217929": "Why trade stocks & indices on Deriv",
