+++ src/pages/markets/components/markets/_commodities.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Loadable from '@loadable/component'
 import AvailableTrades from '../helper/_available-trades'
-import { WhyTrade } from '../sections/_why-trade'
+import { WhyTrade } from '../sections/_trade-details'
 import commodities from '../../static/content/_commodities'
 import { commodities_cfds } from '../../static/content/_cfds'
 import { commodities_options } from '../../static/content/_digital-options'
@@ -11,6 +11,7 @@ import { StyledBox } from '../../static/style/_markets-style'
 import { SimpleStepContentElement } from '../../static/content/_simple_step_content'
 import { Localize, localize } from 'components/localization'
 import useRegion from 'components/hooks/use-region'
+import { FullWidthMultiColumn } from 'components/elements/full-width-multicolumn'
 //Lazy-load
 const SimpleSteps = Loadable(() => import('components/custom/_simple-steps'))
 const OtherMarkets = Loadable(() => import('../sections/_other-markets'))
@@ -27,15 +28,6 @@ const Commodities = ({ simple_step_content }: CommoditiesProps) => {
 
     return (
         <>
-            <WhyTrade header={<Localize translate_text="Why trade commodities on Deriv" />}>
-                {commodities.map((content, index) => (
-                    <StyledBox
-                        key={index}
-                        text={content.text}
-                        icon={<img src={content.src} alt={content.alt} />}
-                    />
-                ))}
-            </WhyTrade>
             {is_eu ? (
                 <AvailableTrades
                     CFDs={<CFDs market_content={commodities_cfds} />}
@@ -57,6 +49,17 @@ const Commodities = ({ simple_step_content }: CommoditiesProps) => {
                     }
                 />
             )}
+            <FullWidthMultiColumn
+                header={<Localize translate_text="Why trade commodities on Deriv" />}
+            >
+                {commodities.map((content, index) => (
+                    <StyledBox
+                        key={index}
+                        text={content.text}
+                        icon={<img src={content.src} alt={content.alt} />}
+                    />
+                ))}
+            </FullWidthMultiColumn>
             <SimpleSteps
                 header={
                     <Localize translate_text="Start trading commodities on Deriv in 3 simple steps" />
