+++ src/pages/markets/components/markets/_forex.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import Loadable from '@loadable/component'
-import { WhyTrade } from '../sections/_why-trade'
 import AvailableTrades from '../helper/_available-trades'
 import { forex_content, forex_content_eu } from '../../static/content/_forex'
 import { forex_cfds, forex_cfds_eu } from '../../static/content/_cfds'
@@ -13,6 +12,7 @@ import { StyledBox } from '../../static/style/_markets-style'
 import { SimpleStepContentElement } from '../../static/content/_simple_step_content'
 import useRegion from 'components/hooks/use-region'
 import { Localize, localize } from 'components/localization'
+import { FullWidthMultiColumn } from 'components/elements/full-width-multicolumn'
 
 //Lazy-load
 const SimpleSteps = Loadable(() => import('components/custom/_simple-steps'))
@@ -25,15 +25,6 @@ const Forex = ({ simple_step_content }: ForexProps) => {
     const { is_row, is_eu } = useRegion()
     return (
         <>
-            <WhyTrade header={<Localize translate_text="Why trade forex on Deriv" />}>
-                {(is_eu ? forex_content_eu : forex_content).map((content, index) => (
-                    <StyledBox
-                        key={index}
-                        text={content.text}
-                        icon={<img src={content.src} alt={content.alt} />}
-                    />
-                ))}
-            </WhyTrade>
             <AvailableTrades
                 CFDs={<CFDs market_content={is_eu ? forex_cfds_eu : forex_cfds} />}
                 DigitalOptions={
@@ -50,6 +41,15 @@ const Forex = ({ simple_step_content }: ForexProps) => {
                 // name="Forex"
                 display_title={<Localize translate_text="Forex trades available on Deriv" />}
             />
+            <FullWidthMultiColumn header={<Localize translate_text="Why trade forex on Deriv" />}>
+                {(is_eu ? forex_content_eu : forex_content).map((content, index) => (
+                    <StyledBox
+                        key={index}
+                        text={content.text}
+                        icon={<img src={content.src} alt={content.alt} />}
+                    />
+                ))}
+            </FullWidthMultiColumn>
             <SimpleSteps
                 header={
                     <Localize translate_text="Start trading forex on Deriv in 3 simple steps" />
