+++ src/pages/markets/components/markets/_stock-indices.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import Loadable from '@loadable/component'
-import { WhyTrade } from '../sections/_why-trade'
 import AvailableTrades from '../helper/_available-trades'
 import stock_content from '../../static/content/_stock'
 import { stock_cfds } from '../../static/content/_cfds'
@@ -10,6 +9,7 @@ import DigitalOptions from '../sub-markets/_digital-options'
 import { StyledBox } from '../../static/style/_markets-style'
 import { SimpleStepContentElement } from '../../static/content/_simple_step_content'
 import { localize, Localize } from 'components/localization'
+import { FullWidthMultiColumn } from 'components/elements/full-width-multicolumn'
 
 //Lazy-load
 const SimpleSteps = Loadable(() => import('components/custom/_simple-steps'))
@@ -26,15 +26,6 @@ const StockIndices = ({ simple_step_content }: StockIndicesProps) => {
 
     return (
         <div>
-            <WhyTrade header={<Localize translate_text="Why trade stocks & indices on Deriv" />}>
-                {stock_content.map((content, index) => (
-                    <StyledBox
-                        key={index}
-                        text={content.text}
-                        icon={<img src={content.src} alt={content.alt} />}
-                    />
-                ))}
-            </WhyTrade>
             <AvailableTrades
                 CFDs={<CFDs market_tab_name={'stock-indices'} market_content={stock_cfds} />}
                 DigitalOptions={
@@ -47,6 +38,17 @@ const StockIndices = ({ simple_step_content }: StockIndicesProps) => {
                     <Localize translate_text="Stocks & indices trades available on Deriv" />
                 }
             />
+            <FullWidthMultiColumn
+                header={<Localize translate_text="Why trade stocks & indices on Deriv" />}
+            >
+                {stock_content.map((content, index) => (
+                    <StyledBox
+                        key={index}
+                        text={content.text}
+                        icon={<img src={content.src} alt={content.alt} />}
+                    />
+                ))}
+            </FullWidthMultiColumn>
             <SimpleSteps
                 header={
                     <Localize translate_text="Start trading stocks & indices on Deriv in 3 simple steps" />
