+++ src/pages/markets/components/markets/_synthetic-indices.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import Loadable from '@loadable/component'
-import { WhyTrade } from '../sections/_why-trade'
 import AvailableTrades from '../helper/_available-trades'
 import synthetic_content from '../../static/content/_synthetic'
 import { synthetic_cfds, synthetic_cfds_eu } from '../../static/content/_cfds'
@@ -11,10 +10,12 @@ import Multipliers from '../sub-markets/_multipliers'
 import DigitalOptions from '../sub-markets/_digital-options'
 import { StyledBox } from '../../static/style/_markets-style'
 import { SimpleStepContentElement } from '../../static/content/_simple_step_content'
+import { TradeDetails } from '../sections/_trade-details'
 import { Localize, localize } from 'components/localization'
 const SimpleSteps = Loadable(() => import('components/custom/_simple-steps'))
 const OtherMarkets = Loadable(() => import('../sections/_other-markets'))
 import useRegion from 'components/hooks/use-region'
+import { FullWidthMultiColumn } from 'components/elements/full-width-multicolumn'
 
 type StockIndicesProps = {
     simple_step_content: SimpleStepContentElement[]
@@ -25,20 +26,11 @@ const StockIndices = ({ simple_step_content }: StockIndicesProps) => {
 
     return (
         <div>
-            <WhyTrade
+            <TradeDetails
                 description={
                     <Localize translate_text="Deriv’s proprietary synthetics simulate real-world market movements. Backed by a cryptographically secure random number generator, these indices are available to trade 24/7 and are unaffected by regular market hours, global events, or market and liquidity risks." />
                 }
-                header={<Localize translate_text="Why trade synthetics on Deriv" />}
-            >
-                {synthetic_content.map((content, index) => (
-                    <StyledBox
-                        key={index}
-                        text={content.text}
-                        icon={<img src={content.src} alt={content.alt} />}
-                    />
-                ))}
-            </WhyTrade>
+            />
             <AvailableTrades
                 CFDs={
                     <CFDs
@@ -60,6 +52,20 @@ const StockIndices = ({ simple_step_content }: StockIndicesProps) => {
                 }
                 display_title={<Localize translate_text="Synthetics trades available on Deriv" />}
             />
+            <FullWidthMultiColumn
+                description={
+                    <Localize translate_text="Deriv’s proprietary synthetics simulate real-world market movements. Backed by a cryptographically secure random number generator, these indices are available to trade 24/7 and are unaffected by regular market hours, global events, or market and liquidity risks." />
+                }
+                header={<Localize translate_text="Why trade synthetics on Deriv" />}
+            >
+                {synthetic_content.map((content, index) => (
+                    <StyledBox
+                        key={index}
+                        text={content.text}
+                        icon={<img src={content.src} alt={content.alt} />}
+                    />
+                ))}
+            </FullWidthMultiColumn>
             <SimpleSteps
                 header={
                     <Localize translate_text="Start trading synthetics on Deriv in 3 simple steps" />
